+++ CHANGES.rst
@@ -17,7 +17,8 @@ dev (master)
 
 * Py3: Use ``ssl.create_default_context()`` when available. (Issue #473)
 
-* Emit ``InsecureRequestWarning`` for *every* insecure request. (Issue #496)
+* Emit ``InsecureRequestWarning`` for *every* insecure HTTPS request.
+  (Issue #496)
 
 * ... [Short description of non-trivial change.] (Issue #)